Changes: App.js is passing shuffleArrays as a prop to the Review page. Review has been set to run shuffle function in componentDidMount. ComponentDidUpdate has been added to Review so the states can be properly updated after receiving props.
Issue: Page is still not loading cards based off the props passed. The page wants to read a not-yet-existing array index which is causing the page to error.

Resolution needed: Do not allow page elements to attempt to access not-yet-existing data and properly update after the state has been updated from props.