Changes: Comments added to Card, LangOptions, Tab, Table, Quiz, and Review. Comments contain information regarding the function of individual pieces of code as well as TODOs for the future.
Issue: DRY could not be completed at this time for entryValidation. Function is used in two areas and should be consolidated into one. It appears the current config wanted to keep the state in App.js for guessInput. It seemed counterproductive to pass a function to child, have the child's child pass data to that function, have the parent process the data, and then pass the outcome down through props to the children. A simpler/more elegant solution would be ideal.